requirements.txt
numpy>=1.24.3
setuptools>=66.0.0
scipy>=1.10.1
matplotlib>=3.7.1
scikit-learn>=1.2.2
tqdm>=4.65.0
pandas>=2.0.1
nose>=1.3.7
pybind11>=2.10.4
pydsge
torch<=2.3.0
requests
six>=1.15.0
optuna
gensim>=4.1.2
python-easygraph >=1.5
torch
torch-sparse
numpy
scipy
metis